pyproject.toml
[project]
name = "trilium-pydantic"
version = "0.1.1"
description = "Type-safe Python library for TriliumNext ETAPI"
authors = [
    {name = "Bullish-Design", email = "BullishDesignEngineering@gmail.com"}
]
dependencies = [
    "trilium-py>=1.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
  #"mypy>=1.0.0",
  #"ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/trilium_pydantic", "src/tests", "src/examples"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

#[tool.ruff]
#line-length = 79
#target-version = "py311"

#[tool.ruff.lint]
#select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "B", "A", "C4"]
#ignore = []

#[tool.mypy]
#python_version = "3.11"
#strict = true
#warn_return_any = true
#warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/trilium_pydantic --cov-report=term-missing"
